Use realistic numbers and improve display scaling

+++ src/visualizer/types.py
@@ -1,4 +1,4 @@
-ScreenDimension = int
+ScreenDimension = float
 Fps = int
 TimePerFrame = float
 SpeedScale = float
